dry up auth notification
