+++ components/NativeApp/AppSignInOverlay.js
@@ -1,102 +1,18 @@
-import React, { useEffect, useContext, useState } from 'react'
+import React from 'react'
 import {
   Overlay,
   OverlayToolbar,
   OverlayToolbarConfirm,
   Interaction,
   OverlayBody,
-  Button,
   useColorContext
 } from '@project-r/styleguide'
-import TokenAuthorization from '../Auth/TokenAuthorization'
-import { DEFAULT_TOKEN_TYPE } from '../constants'
-import * as base64u from '../../lib/utils/base64u'
-import isEmail from 'validator/lib/isEmail'
 import { MdClose } from 'react-icons/md'
 import withT from '../../lib/withT'
-import RawHtmlTranslation from '../RawHtmlTranslation'
-
-const { H1, P } = Interaction
-
-const knownTypes = [
-  'email-confirmed',
-  'invalid-email',
-  'invalid-token',
-  'session-denied',
-  'token-authorization',
-  'unavailable'
-]
+import AuthNotification from '../Auth/Notification'
 
 const AppSignInOverlay = ({ onClose, query, setQuery, t }) => {
   const [colorScheme] = useColorContext()
-  const { context, token, tokenType, noAutoAuthorize } = query
-  let { type, email } = query
-  if (email !== undefined) {
-    try {
-      if (base64u.match(email)) {
-        email = base64u.decode(email)
-      }
-    } catch (e) {}
-
-    if (!isEmail(email)) {
-      type = 'invalid-email'
-      email = ''
-    }
-  }
-
-  let title
-  let content
-  let isUnkownType
-
-  if (type === 'token-authorization') {
-    content = (
-      <TokenAuthorization
-        email={email}
-        token={token}
-        tokenType={tokenType || DEFAULT_TOKEN_TYPE}
-        noAutoAuthorize={noAutoAuthorize}
-        context={context}
-        goTo={(type, email, context) => {
-          if (type === 'email-confirmed') {
-            onClose()
-            return
-          }
-          setQuery({ type, email, context })
-        }}
-        onCloseAuthorization={onClose}
-      />
-    )
-  } else {
-    title = t.first(
-      [`notifications/${type}/${context}/title`, `notifications/${type}/title`],
-      undefined,
-      ''
-    )
-    if (!title && !knownTypes.includes(type)) {
-      title = t('notifications/unkown/title')
-      isUnkownType = true
-    }
-
-    content = (
-      <>
-        <P>
-          <RawHtmlTranslation
-            first={[
-              `notifications/${type}/${context}/text`,
-              `notifications/${type}/text`
-            ]}
-            replacements={query}
-            missingValue={isUnkownType ? t('notifications/unkown/text') : ''}
-          />
-        </P>
-        <br />
-        <br />
-        <Button block primary onClick={onClose}>
-          {t('notifications/closeButton/app')}
-        </Button>
-      </>
-    )
-  }
 
   return (
     <Overlay onClose={onClose}>
@@ -110,13 +26,17 @@ const AppSignInOverlay = ({ onClose, query, setQuery, t }) => {
         />
       </OverlayToolbar>
       <OverlayBody>
-        {title && (
-          <>
-            <H1>{title}</H1>
-            <br />
-          </>
-        )}
-        {content}
+        <AuthNotification
+          query={query}
+          onClose={onClose}
+          goTo={(type, email, context) => {
+            if (type === 'email-confirmed') {
+              onClose()
+              return
+            }
+            setQuery({ type, email, context })
+          }}
+        />
       </OverlayBody>
     </Overlay>
   )
